requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
antlr4-python3-runtime==4.8
    # via hydra-core
apscheduler==3.6.3
    # via python-telegram-bot
certifi==2020.12.5
    # via
    #   python-telegram-bot
    #   requests
chardet==4.0.0
    # via requests
hydra-core==1.0.6
    # via -r requirements.in
idna==2.10
    # via requests
importlib-resources==5.1.4
    # via hydra-core
omegaconf==2.0.6
    # via
    #   -r requirements.in
    #   hydra-core
python-telegram-bot==13.5
    # via -r requirements.in
pytz==2021.1
    # via
    #   apscheduler
    #   python-telegram-bot
pyyaml==5.4.1
    # via omegaconf
requests==2.25.1
    # via -r requirements.in
six==1.16.0
    # via apscheduler
tornado==6.1
    # via python-telegram-bot
typing-extensions==3.10.0.0
    # via omegaconf
tzlocal==2.1
    # via apscheduler
urllib3==1.26.5
    # via requests
zipp==3.4.1
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
